+++ rs-lib/src/network/com_interfaces/mod.rs
@@ -4,11 +4,8 @@ pub mod base_interface;
 pub mod matchbox_js_interface;
 #[cfg(feature = "wasm_serial")]
 pub mod serial_js_interface;
-
 #[cfg(feature = "wasm_webrtc")]
 pub mod webrtc_js_interface;
-#[cfg(feature = "wasm_webrtc")]
-pub mod webrtc_js_interface_new;
 #[cfg(feature = "wasm_websocket_client")]
 pub mod websocket_client_js_interface;
 #[cfg(feature = "wasm_websocket_server")]
